moved post functionality to its own component
